restructure to allow easy adding of generators
