+++ python/quick_der/generators/source.pyi
@@ -1,0 +1,41 @@
+# Stubs for quick_der.generators.source (Python 3.6)
+#
+# NOTE: This dynamically typed stub was automatically generated by stubgen.
+
+from quick_der.generators import QuickDERgeneric
+from typing import Any
+
+logger: Any
+
+class QuickDER2source(QuickDERgeneric):
+    to_be_defined: Any = ...
+    to_be_overlaid: Any = ...
+    cursor_offset: Any = ...
+    nested_typerefs: Any = ...
+    nested_typecuts: Any = ...
+    semamod: Any = ...
+    refmods: Any = ...
+    buffer: Any = ...
+    linebuffer: Any = ...
+    comma1: Any = ...
+    comma0: Any = ...
+    issued_typedefs: Any = ...
+    pack_funmap: Any = ...
+    def __init__(self, semamod, outfn, refmods) -> None: ...
+    def write(self, txt): ...
+    def writeln(self, txt: str = ...): ...
+    def close(self): ...
+    def generate_head(self): ...
+    def generate_tail(self): ...
+    def generate_unpack(self): ...
+    def generate_pack(self): ...
+    def generate_pack_node(self, node, tp, fld): ...
+    def packValueAssignment(self, node, tp, fld): ...
+    def packDefinedType(self, node, tp, fld): ...
+    def packSimpleType(self, node, tp, fld): ...
+    def packTypeAssignment(self, node, tp, fld): ...
+    def packSequenceType(self, node, tp, fld, naked: bool = ...): ...
+    def packSetType(self, node, tp, fld, naked: bool = ...): ...
+    def packChoiceType(self, node, tp, fld, naked: bool = ...): ...
+    def packRepeatingStructureType(self, node, tp, fld): ...
+    def packTaggedType(self, node, tp, fld): ...
