+++ python/quick_der/util.pyi
@@ -1,0 +1,10 @@
+from typing import Any
+
+class dprint:
+    enable: bool = ...
+    def __init__(self, s, *args) -> None: ...
+
+def tosym(name): ...
+
+api_prefix: str
+dertag2atomsubclass: Any
